+++ src/index.js
@@ -1,5 +1,6 @@
 require('babel-polyfill');
 
+const clc = require('cli-color');
 const _path = require('path');
 const axios = require('axios');
 const sanitize = require('sanitize-filename');
@@ -7,22 +8,36 @@ const _ = require('lodash');
 const { flatMap } = require('lodash');
 const fs = require('fs');
 const mkdirp = require('mkdirp');
+const { stringReplace, parseUrl } = require('./utils');
 
 // get root project directory
 const appRootDir = _path.resolve('.').split('/node_modules')[0];
+
 class easyApiFixtures {
-  constructor(configPath = _path.join(appRootDir, 'fixtures.config.js')) {
+  constructor(configPath, feedback = false) {
+    this.requests = [];
+    this.emittedFiles = [];
+    this.configPath =
+      configPath || _path.join(appRootDir, 'fixtures.config.js');
     this.appRootDir = appRootDir;
+    this.feedback = feedback;
     this.defaults = {
+      requestFunction: axios.get,
       output: {
         filename: '[name].json',
         uglified: false,
-        versionInPath: true,
-        endpointInPath: true,
-        aliasInPath: true,
       },
     };
-    this.config = this.parseConfig(this.constructor.loadFile(configPath));
+    try {
+      if (this.feedback) console.log(`\nLoading config: ${this.configPath}`);
+      this.config = this.parseConfig(
+        this.constructor.loadFile(this.configPath)
+      );
+      if (this.feedback)
+        console.log(clc.green('\nSuccessfully loaded configuration.'));
+    } catch (err) {
+      throw err;
+    }
   }
 
   /**
@@ -43,15 +58,29 @@ class easyApiFixtures {
    * @param {Object} api - An object that holds information on an api from the config file
    * @param {Function} requestFn - (default axios) the promise based http request function
    */
-  static async getFixturesDataFromApi(api, requestFn = axios.get) {
-    const _getData = (url, endpoint, slug) =>
-      requestFn(`${url}/${endpoint}/${slug}`)
-        .then(result => ({
-          slug,
-          endpoint,
-          data: result.data,
-        }))
-        .catch(err => console.log(err));
+  async getFixturesDataFromApi(api, requestFn = this.config.requestFunction) {
+    const _getData = (url, endpoint, slug) => {
+      try {
+        return requestFn(`${url}/${endpoint}/${slug}`)
+          .then(result => {
+            if (this.feedback)
+              console.log(
+                clc.green('\t[success] '),
+                `${url}/${endpoint}/${slug}`
+              );
+            this.requests.push(`${url}/${endpoint}/${slug}`);
+            return { slug, endpoint, data: result.data };
+          })
+          .catch(
+            () =>
+              this.feedback &&
+              console.log(clc.red('\t[failed] '), `${url}/${endpoint}/${slug}`)
+          );
+      } catch (err) {
+        console.log(clc.red(err));
+      }
+    };
+
     const fixtures = await Promise.all(
       _.flatMap(api.fixture, fixture => {
         const { url } = api;
@@ -59,7 +88,7 @@ class easyApiFixtures {
           _.map(fixture.slug, slug => _getData(url, endpoint, slug))
         );
       })
-    );
+    ).catch(err => console.log(clc.red(err)));
     return fixtures;
   }
 
@@ -67,8 +96,10 @@ class easyApiFixtures {
    * Tests for directory's existence and creates it if it does not
    * @param {String} filePath - path to directory
    */
-  static async ensureDirectoryExistence(filePath) {
+  async ensureDirectoryExistence(filePath) {
     if (!fs.existsSync(filePath)) {
+      if (this.feedback)
+        console.log(clc.yellow(`${filePath} did not exist...creating it.`));
       mkdirp.sync(filePath);
     }
   }
@@ -80,9 +111,14 @@ class easyApiFixtures {
   parseConfig(config) {
     const newConfig = config;
     newConfig.output = Object.assign({}, this.defaults.output, config.output);
+    newConfig.requestFunction = _.get(config, 'requestFunction', axios.get);
     const apiArray = flatMap([newConfig.api]);
     const apis = apiArray.map(api =>
       Object.assign({}, api, {
+        _pathMap: {
+          '[api]': sanitize(_.get(api, 'alias', _.get(api, 'url', ''))),
+          '[version]': _.get(api, 'version', ''),
+        },
         alias: sanitize(api.alias || api.url),
         fixture: flatMap([api.fixture]).map(fixture =>
           Object.assign({}, fixture, {
@@ -100,13 +136,10 @@ class easyApiFixtures {
    * @param {Object} api - An object that holds information on an api from the config file
    */
   getBasePath(api) {
-    const { output } = this.config;
-    return _path.join(
-      this.appRootDir,
-      output.path,
-      output.aliasInPath ? _.get(api, 'alias', '') : '',
-      output.versionInPath ? _.get(api, 'version', '') : ''
-    );
+    const { path } = this.config.output;
+    const pathToUse = api.path || path;
+
+    return _path.join(this.appRootDir, stringReplace(pathToUse, api._pathMap));
   }
 
   /**
@@ -121,7 +154,14 @@ class easyApiFixtures {
    * @param {String} slug - the slug of the request
    */
   getFileName(slug) {
-    const name = this.config.output.filename.replace('[name]', slug);
+    const replaceObj = {
+      '?': '_where_',
+      '=': '_equals_',
+      '&': '_and_',
+      ' ': '_',
+    };
+    const parsedSlug = stringReplace(slug, replaceObj);
+    const name = this.config.output.filename.replace('[name]', parsedSlug);
     return name;
   }
 
@@ -130,18 +170,15 @@ class easyApiFixtures {
    * @param {String} target - the real URL to mock
    */
   request(target) {
-    const regEx = /.*:\/\/.*?(?=\/)|\[mock\]/gm;
-    const base = target.match(regEx)[0];
-    const endpoint = target.split(regEx)[1];
-    const { endpointInPath } = this.config.output;
-    const pathArray = endpoint.split('/').filter(a => a);
-    const fixturePath = _path.join(
-      this.getBasePath(
-        this.config.api.filter(api => api.url.includes(base))[0]
-      ),
-      endpointInPath ? pathArray[0] : '',
-      this.getFileName(pathArray[1])
+    const { path, base } = parseUrl(target);
+    console.log(base);
+    const [endpoint, filename] = path.split('/').filter(a => a);
+    console.log({ endpoint, filename });
+    const basePath = this.getBasePath(
+      this.config.api.filter(api => api.url.includes(base))[0]
     );
+    const updatedPath = stringReplace(basePath, { '[endpoint]': endpoint });
+    const fixturePath = _path.join(updatedPath, this.getFileName(filename));
     return this.constructor.loadFile(fixturePath);
   }
 
@@ -152,10 +189,14 @@ class easyApiFixtures {
     this.fixtureData = await Promise.all(
       this.config.api.map(async api => {
         const fixturePath = this.getBasePath(api);
-        return {
-          fixturePath,
-          fixtures: await this.constructor.getFixturesDataFromApi(api),
-        };
+        try {
+          return {
+            fixturePath,
+            fixtures: await this.getFixturesDataFromApi(api),
+          };
+        } catch (err) {
+          throw err;
+        }
       })
     );
   }
@@ -165,28 +206,45 @@ class easyApiFixtures {
    * @param {Object} {fixturePath, slug, endpoint, data}
    */
   async writeFile({ fixturePath, slug, endpoint, data }) {
-    await this.constructor.ensureDirectoryExistence(
-      _path.join(fixturePath, endpoint)
-    );
-    const file = _path.join(fixturePath, endpoint, this.getFileName(slug));
-    fs.writeFile(file, JSON.stringify(data, null, 2), err => {
-      if (err) {
-        console.error(err);
-        return;
-      }
-      console.log('\x1b[32m', file, 'has been created', '\x1b[0m');
-    });
+    if (data && slug && endpoint && fixturePath) {
+      const path = stringReplace(fixturePath, { '[endpoint]': endpoint });
+      await this.ensureDirectoryExistence(path);
+      const filename = this.getFileName(slug);
+      const file = _path.join(path, filename);
+      await fs.writeFile(file, JSON.stringify(data, null, 2), async err => {
+        if (err) {
+          console.error(clc.red('\t[failed] '), filename);
+          return;
+        }
+        await this.emittedFiles.push(filename);
+        console.log(clc.green('\t[success] '), filename);
+      });
+    }
   }
 
   /**
    * Runs the application to create fixtures
    */
   async run() {
-    await this.loadData();
-    this.fixtureData.forEach(data => {
+    if (this.feedback)
+      console.log(
+        '\nRetrieving data from APIs...',
+        clc.bgWhite.black('\n\n\t STATUS        REQUEST        ')
+      );
+    try {
+      await this.loadData();
+    } catch (err) {
+      throw err;
+    }
+    if (this.feedback)
+      console.log(
+        '\nWriting API data to files...',
+        clc.bgWhite.black('\n\n\t STATUS        OUTPUT        ')
+      );
+    await this.fixtureData.forEach(async data => {
       const { fixturePath } = data;
-      data.fixtures.forEach(fixture => {
-        this.writeFile({ fixturePath, ...fixture });
+      await data.fixtures.forEach(async fixture => {
+        await this.writeFile({ fixturePath, ...fixture });
       });
     });
   }