+++ src/run.js
@@ -1,7 +1,16 @@
 #! /usr/bin/env node
-
+const clc = require('cli-color');
+const packageJSON = require('../package.json');
 const EasyApiFixtures = require('./index');
 
-const easyApiFixtures = new EasyApiFixtures();
+const easyApiFixtures = new EasyApiFixtures(null, true);
 
-easyApiFixtures.run();
+const run = async () => {
+  console.log(
+    clc.black.bgBlue(`\n    EASY-API-FIXTURES ${packageJSON.version}    `)
+  );
+
+  easyApiFixtures.run();
+};
+
+run();