+++ test/fixtures/mock.config.js
@@ -1,18 +1,44 @@
 module.exports = {
   output: {
     filename: '[name].json',
-    path: 'test/fixtures',
+    path: 'test/fixtures/[version]/[endpoint]',
     aliasInPath: false,
   },
-  api: {
-    url: '[mock]',
-    version: 'v1',
-    alias: 'mock',
-    fixture: [
-      {
-        endpoint: 'mock',
-        slug: 'data',
-      },
-    ],
-  },
+  api: [
+    {
+      url: 'mock-api',
+      version: 'v1',
+      alias: 'mock',
+      fixture: [
+        {
+          endpoint: 'mock',
+          slug: 'data',
+        },
+      ],
+    },
+    {
+      url: 'mock-api',
+      version: 'v1',
+      alias: 'mock',
+      path: 'test/fixtures/[api]/[endpoint]',
+      fixture: [
+        {
+          endpoint: 'mock',
+          slug: 'data',
+        },
+      ],
+    },
+    {
+      url: 'mock-api',
+      version: 'v1',
+      alias: 'mock',
+      path: 'test/fixtures/[api]/[version]/[endpoint]',
+      fixture: [
+        {
+          endpoint: 'mock',
+          slug: 'data',
+        },
+      ],
+    },
+  ],
 };